+++ util/xheep_gen/peripherals/abstractions.py
@@ -3,7 +3,6 @@
 import os.path as path
 from abc import (
     ABC,
-    abstractmethod,
 )  # Used to define abstract classes that cannot be instantiated, only well defined subclasses can be instantiated.
 from enum import Enum
 from copy import deepcopy
@@ -217,42 +216,62 @@ class UserPeripheral(Peripheral, ABC):
     """
 
 
-class PeripheralDomain(ABC):
+class PeripheralDomain:
     """
-    Abstract class representing a peripheral domain. This class cannot be instantiated.
+    A peripheral domain: a group of peripherals connected to the system bus as
+    an independent node. Each domain can be assigned to a power domain and can
+    support clock gating, so that domains can be grouped and switched off or
+    clock gated together.
 
     :param str name: The name of the peripheral domain. Convention : starts with a capital letter and is in singular form (no "peripheral domain" at the end)
     :param int start_address: The start address of the peripheral domain.
     :param int length: The length of the peripheral domain.
+    :param str power_domain: The name of the power domain the domain belongs to. `None` means always-on. Domains sharing the same power domain name are switched on/off together.
+    :param bool clock_gating: `True` if the domain supports clock gating.
     :param list[Peripheral] peripherals: The list of peripherals in the domain. There can be more than one instance of the same peripheral.
     """
 
     _name: str
     _start_address: int
     _length: int
+    _power_domain: Optional[str]
+    _clock_gating: bool
     _peripherals: List[
         Peripheral
     ]  # type has to be precised for filtering in validation
 
-    @abstractmethod
     def __init__(
         self,
         name: str,
         start_address: int,
         length: int,
+        power_domain: Optional[str] = None,
+        clock_gating: bool = False,
         peripherals: Optional[List[Peripheral]] = None,
     ):
         """
-        Initialize the peripheral domain. Is abstract because each peripheral domain has its own way of initializing without letting the user define start address and length.
+        Initialize the peripheral domain.
 
         :param str name: The name of the peripheral domain. Convention : starts with a capital letter and is in singular form (no "peripheral domain" at the end)
         :param int start_address: The start address of the peripheral domain.
         :param int length: The length of the peripheral domain.
+        :param str power_domain: The name of the power domain the domain belongs to. `None` means always-on.
+        :param bool clock_gating: `True` if the domain supports clock gating.
         :param list[Peripheral] peripherals: Optional initial list of peripherals.
         """
+        if power_domain is not None and type(power_domain) is not str:
+            raise TypeError(
+                f"PeripheralDomain.power_domain should be of type str not {type(power_domain)}"
+            )
+        if type(clock_gating) is not bool:
+            raise TypeError(
+                f"PeripheralDomain.clock_gating should be of type bool not {type(clock_gating)}"
+            )
         self._name = f"{name} Peripheral Domain"
         self._start_address = start_address
         self._length = length
+        self._power_domain = power_domain
+        self._clock_gating = clock_gating
         self._peripherals = []
         if peripherals is not None:
             if type(peripherals) is not list:
@@ -260,23 +279,97 @@ class PeripheralDomain(ABC):
             for peripheral in peripherals:
                 self.add_peripheral(peripheral)
 
-    @abstractmethod
+    def get_name(self):
+        """
+        :return: The name of the peripheral domain.
+        :rtype: str
+        """
+        return self._name
+
     def add_peripheral(self, peripheral: Peripheral):
         """
-        Add a peripheral to the domain. The peripheral should be fully configured when added. If the peripheral has no offset, it will be automatically computed during build. Must be defined by the subclass.
+        Add a peripheral to the domain. The peripheral should be fully
+        configured when added. If the peripheral has no offset, it will be
+        automatically computed during build.
 
         :param Peripheral peripheral: The peripheral to add.
+        :raise ValueError: when peripheral is not a Peripheral.
         """
-        ...
+        if not isinstance(peripheral, Peripheral):
+            raise ValueError("Peripheral is not a Peripheral")
+        self._peripherals.append(peripheral)
 
-    @abstractmethod
     def remove_peripheral(self, peripheral: Peripheral):
         """
-        Remove a peripheral from the domain. Must be defined by the subclass.
+        Remove a peripheral from the domain.
 
         :param Peripheral peripheral: The peripheral to remove.
         """
-        ...
+        if peripheral not in self._peripherals:
+            print(
+                f"Warning : Peripheral {peripheral.get_name()} is not in the domain {self._name}"
+            )
+            return
+        self._peripherals.remove(peripheral)
+
+    # ------------------------------------------------------------
+    # Power / Clock-Gating Domain
+    # ------------------------------------------------------------
+
+    def set_power_domain(self, power_domain: Optional[str]):
+        """
+        Assign the domain to a power domain. Domains sharing the same power
+        domain name are switched on/off together.
+
+        :param str power_domain: The name of the power domain. `None` means always-on.
+        :raise TypeError: when power_domain is of incorrect type.
+        """
+        if power_domain is not None and type(power_domain) is not str:
+            raise TypeError(
+                f"PeripheralDomain.power_domain should be of type str not {type(power_domain)}"
+            )
+        self._power_domain = power_domain
+
+    def get_power_domain(self):
+        """
+        :return: The name of the power domain the domain belongs to, `None` if always-on.
+        :rtype: str
+        """
+        return self._power_domain
+
+    def has_power_domain(self) -> bool:
+        """
+        :return: `True` if the domain belongs to a switchable power domain, `False` if always-on.
+        :rtype: bool
+        """
+        return self._power_domain is not None
+
+    def is_always_on(self) -> bool:
+        """
+        :return: `True` if the domain is always-on (no switchable power domain).
+        :rtype: bool
+        """
+        return self._power_domain is None
+
+    def set_clock_gating(self, clock_gating: bool):
+        """
+        Enable or disable clock gating support for the domain.
+
+        :param bool clock_gating: `True` if the domain supports clock gating.
+        :raise TypeError: when clock_gating is of incorrect type.
+        """
+        if type(clock_gating) is not bool:
+            raise TypeError(
+                f"PeripheralDomain.clock_gating should be of type bool not {type(clock_gating)}"
+            )
+        self._clock_gating = clock_gating
+
+    def has_clock_gating(self) -> bool:
+        """
+        :return: `True` if the domain supports clock gating.
+        :rtype: bool
+        """
+        return self._clock_gating
 
     def get_start_address(self):
         """
